Update Funky.html
OK! It is ~midnight and I am still stuck on trying to get the "const funkyRotate" function to actually call on the right aspect of the new funkyList var. Gonna be committing the new Java and CSS with this to show how rough I made this. Currently, selecting a button only moves the first row, instead of the row related to the button. With that in mind, I am unsure on how to actually point to a specific section in the funkyList array for the flexBox const (which, in turn, leads to the function just always defaulting to the first row, instead of the row indicated by the array #). Plz halp );